requirements.txt
frontend==0.0.3
PyMuPDF==1.26.3
Pillow==11.3.0